動いた！

+++ F446-main/src/sts3032.cpp
@@ -2,7 +2,7 @@
 
 #define SPEED 50  // 速度(0~100%)
 
-void STS3032::Servo_Speed(int currentspeed) {
+void STS3032::Servo_Speed(unsigned long currentspeed) {
     currentspeed = SPEED * 100;
     st.WriteSpe(1,currentspeed,0);
 }
